Server/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0

paddleocr==3.4.0

python-dotenv==1.1.0
pydantic==2.12.5
python-multipart==0.0.22

numpy==2.4.3
pillow==12.1.1
orjson==3.11.7

prometheus-client==0.24.1

ultralytics>=8.3.0

pytest==9.0.2
httpx==0.28.1